Move registry and processors to own folder
