+++ tests/unit/HandlersRegistryTest.php
@@ -17,8 +17,8 @@ use Brain\Monkey\Actions;
 use Inpsyde\Wonolog\Channels;
 use Inpsyde\Wonolog\DefaultHandler;
 use Inpsyde\Wonolog\Factory;
-use Inpsyde\Wonolog\HandlersRegistry;
-use Inpsyde\Wonolog\ProcessorsRegistry;
+use Inpsyde\Wonolog\Registry\HandlersRegistry;
+use Inpsyde\Wonolog\Registry\ProcessorsRegistry;
 use Inpsyde\Wonolog\Tests\UnitTestCase;
 use Monolog\Handler\HandlerInterface;
 use Monolog\Handler\TestHandler;
